New pixel art shader.

+++ src/svarog/render/Camera.java
@@ -29,12 +29,12 @@ public class Camera {
 		return projection.translate(position, new Matrix4f());
 	}
 	
-	public void setProjection(int width, int height, Window window, int scale, int worldWidth, int worldHeight) {
+	public void setProjection(int width, int height, Window window, float scale, int worldWidth, int worldHeight) {
 		projection = new Matrix4f().setOrtho2D(-width/2, width/2, -height/2, height/2);
 		
 		
-		int widthOffset = (window.getWidth()-worldWidth*(scale/2)*4)/2;		// This set world to center of the window
-		int heightOffset = (window.getHeight()-worldHeight*(scale/2)*4)/2;
+		int widthOffset = (int)(window.getWidth()-worldWidth*(scale/2)*4)/2;		// This set world to center of the window
+		int heightOffset = (int)(window.getHeight()-worldHeight*(scale/2)*4)/2;
 		
 		Vector3f offset = new Vector3f(0, 0, 0);
 		if(widthOffset > 0)
